Add files via upload
This was taken from a research paper I had found on SSRN. I modified it quite a bit to only take what I felt would be the important pieces. It calculates directional changes that are above or below a specific threshold that would then indicate a trend move. I am working on making it into a proper function and then adding another function to plot the signals. Uses Pandas DataFrame to store the signals, floats and strings. Can't quite plot strings, but can plot arrows for BUY and SELL from the floats.